Major refactoring in the MVC structure of Central: View and Controller no longer hold references to their sub-elements.
As a sub-view is closed the main 'Model' will be notified and the reference to the sub-Model will be removed. The sub-Controller should be garbage collected automatically as the corresponding sub-Model and sub-View are no longer referenced and thus no references to their sub-Controller exists.

The rest of the open close operations have been added as well.